# app/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SDJR API",
        "description": "API documentation for SDJR application",
        "contact": {
            "email": "admin@sdjr.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/neighborhoods": {
            "get": {
                "tags": [
                    "Neighborhoods"
                ],
                "summary": "List neighborhoods",
                "description": "Get paginated list of neighborhoods",
                "operationId": "indexNeighborhoods",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Neighborhoods"
                ],
                "summary": "Create neighborhood",
                "description": "Store a new neighborhood",
                "operationId": "storeNeighborhood",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NeighborhoodResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/neighborhoods/{id}": {
            "get": {
                "tags": [
                    "Neighborhoods"
                ],
                "summary": "Show neighborhood",
                "description": "Get a specific neighborhood",
                "operationId": "showNeighborhood",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NeighborhoodResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Neighborhoods"
                ],
                "summary": "Update neighborhood",
                "description": "Update a specific neighborhood",
                "operationId": "updateNeighborhood",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NeighborhoodResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Neighborhoods"
                ],
                "summary": "Delete neighborhood",
                "description": "Delete a specific neighborhood",
                "operationId": "deleteNeighborhood",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get all permissions",
                "description": "Returns a list of all permissions created in the system.",
                "operationId": "getPermissions",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PermissionResource"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Create a new permission",
                "description": "Creates a new permission.",
                "operationId": "storePermission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permission created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "users.create"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Permite crear usuarios"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-12-15T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-12-15T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CityResource": {
                "title": "City Resource",
                "description": "City resource representation",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "City ID",
                        "type": "integer",
                        "example": 1
                    },
                    "department_id": {
                        "title": "department_id",
                        "description": "Department ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "City name",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "Bogota"
                    },
                    "status": {
                        "title": "status",
                        "description": "City status",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "A"
                    },
                    "department": {
                        "$ref": "#/components/schemas/DepartmentResource"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Created at timestamp",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Updated at timestamp",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CityResource"
                }
            },
...
            "CountryResource": {
                "title": "Country Resource",
                "description": "Country resource representation",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Country ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Country name",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "Colombia"
                    },
                    "status": {
                        "title": "status",
                        "description": "Country status",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "A"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Created at timestamp",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Updated at timestamp",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CountryResource"
                }
            },
            "DepartmentResource": {
                "title": "Department Resource",
                "description": "Department resource representation",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Department ID",
                        "type": "integer",
                        "example": 1
                    },
                    "country_id": {
                        "title": "country_id",
                        "description": "Country ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Department name",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "Cundinamarca"
                    },
                    "status": {
                        "title": "status",
                        "description": "Department status",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "A"
                    },
                    "country": {
                        "$ref": "#/components/schemas/CountryResource"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Created at timestamp",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Updated at timestamp",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DepartmentResource"
                }
            },
...
            "NeighborhoodResource": {
                "title": "Neighborhood Resource",
                "description": "Neighborhood resource representation",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Neighborhood ID",
                        "type": "integer",
                        "example": 1
                    },
                    "city_id": {
                        "title": "city_id",
                        "description": "City ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Neighborhood name",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "Chapinero"
                    },
                    "status": {
                        "title": "status",
                        "description": "Neighborhood status",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "A"
                    },
                    "city": {
                        "$ref": "#/components/schemas/CityResource"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Created at timestamp",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Updated at timestamp",
                        "type": "array",
                        "items": {
                            "type": "mixed"
                        },
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NeighborhoodResource"
                }
            },
            "PermissionResource": {
                "title": "Permission Resource",
                "description": "Permission resource response",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "users.create"
                    },
                    "description": {
                        "type": "string",
                        "example": "Permite crear usuarios"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}